Fixed style: changed exceptions, tasks and CommandType enum to packages and java coding conventions in Bartholomew.java
