+++ text-ui-test/EXPECTED.TXT
@@ -89,6 +89,43 @@ The event format is incorrect. Please use: event <description> /from <start> /to
 ____________________________________________________________
 
 ____________________________________________________________
+1.[T][ ] read book
+2.[D][ ] return book (by: June 6th)
+3.[E][ ] project meeting (from: Aug 6th 2pm to: 4pm)
+4.[T][X] join sports club
+5.[T][ ] borrow book
+____________________________________________________________
+
+____________________________________________________________
+Noted. I've removed this task:
+  [D][ ] return book (by: June 6th)
+Now you have 5 in the list.
+____________________________________________________________
+
+____________________________________________________________
+Task 100 does not exist in the list.
+____________________________________________________________
+
+____________________________________________________________
+Task 0 does not exist in the list.
+____________________________________________________________
+
+____________________________________________________________
+Task 5 does not exist in the list.
+____________________________________________________________
+
+____________________________________________________________
+"sadasd" is not a valid task number.
+____________________________________________________________
+
+____________________________________________________________
+1.[T][ ] read book
+2.[E][ ] project meeting (from: Aug 6th 2pm to: 4pm)
+3.[T][X] join sports club
+4.[T][ ] borrow book
+____________________________________________________________
+
+____________________________________________________________
 Bye. Hope to see you again soon!
 ____________________________________________________________
